requirements.txt
opencv-python>=4.8.0
numpy>=1.21.0
scikit-image>=0.19.0
matplotlib>=3.5.0
pandas>=1.4.0
seaborn>=0.11.0
scipy>=1.8.0 >> requirements.txt
pillow>=9.0.0
jupyter>=1.0.0
tabulate>=0.9.0